+++ docs/CHANGELOG.md
@@ -1,3 +1,7 @@
+3.20.0 / 2021-09-10
+  * Use strnum npm package to parse string to number
+    * breaking change: long number will be parsed to scientific notation.
+
 3.19.0 / 2021-03-14
   * License changed to MIT original
   * Fix #321 : namespace tag parsing